+++ src/moonlink_connectors/src/error.rs
@@ -79,3 +79,14 @@ impl From<std::io::Error> for Error {
         }
     }
 }
+
+impl<T> From<tokio::sync::mpsc::error::SendError<T>> for Error {
+    fn from(err: tokio::sync::mpsc::error::SendError<T>) -> Self {
+        Error::Io {
+            source: Arc::new(std::io::Error::new(
+                std::io::ErrorKind::Other,
+                format!("Channel send error: {}", err),
+            )),
+        }
+    }
+}
